Change to long integer for PSF data stacks

+++ MHD/C_src/CORE_C/quicksort_c.h
@@ -10,7 +10,7 @@ void quicksort_int_c(int *ivec, int *list, int lo, int hi);
 void quicksort_double_c(double *dvec, int *list, int lo, int hi);
 void quicksort_real_c(float *rvec, int *list, int lo, int hi);
 
-void quicksort_double_2idx_c(double *dvec, int *list1, int *list2, int lo, int hi);
+void quicksort_double_2idx_c(double *dvec, int *list1, int *list2, long lo, long hi);
 
 #endif
